requirements.txt
snowflake-connector-python
snowflake-snowpark-python
pyarrow
pandas
pyspark
python-dotenv
jupyterlab
